+++ great_expectations/execution_engine/sqlalchemy_execution_engine.py
@@ -1,6 +1,5 @@
 import copy
 import datetime
-import enum
 import logging
 import random
 import string
@@ -164,23 +163,6 @@ try:
     import teradatasqlalchemy.types as teradatatypes
 except ImportError:
     teradatasqlalchemy = None
-
-
-class GESqlDialect(enum.Enum):
-    POSTGRESQL = "postgresql"
-    BIGQUERY = "bigquery"
-    DREMIO = "dremio"
-    ORACLE = "oracle"
-    MSSQL = "mssql"
-    MYSQL = "mysql"
-    AWSATHENA = "awsathena"
-    TERADATASQL = "teradatasql"
-    SNOWFLAKE = "snowflake"
-
-    @classmethod
-    def get_all_dialect_names(cls):
-        """Get dialect names for all SQL dialects."""
-        return [dialect_name.value for dialect_name in cls]
 
 
 def _get_dialect_type_module(dialect):
