+++ tests/execution_engine/split_and_sample/test_sqlalchemy_execution_engine_sampling.py
@@ -6,7 +6,7 @@ from great_expectations.core.id_dict import BatchSpec
 from great_expectations.execution_engine.split_and_sample.sqlalchemy_data_sampler import (
     SqlAlchemyDataSampler,
 )
-from great_expectations.execution_engine.sqlalchemy_execution_engine import GESqlDialect
+from great_expectations.execution_engine.sqlalchemy_dialect import GESqlDialect
 
 
 @pytest.mark.parametrize(
